# Conflicts:
#   backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/datagenerator/MerchandiseDataGenerator.java
#   frontend/src/app/components/my-profile/my-profile-tabs/my-invoices-tab/invoice-details/invoice-details.component.html
#   frontend/src/app/components/my-profile/my-profile-tabs/my-invoices-tab/my-invoices-tab.component.html

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/Merchandise.java
@@ -44,12 +44,12 @@ public class Merchandise {
 
     @NotNull
     @Column(nullable = false, name = "price")
-    private int price;
+    private Double price;
 
 
     @NotNull
     @Column(nullable = false, name = "premium_price")
-    private int premiumPrice;
+    private Long premiumPrice;
 
     @NotNull
     @Column(nullable = false, name = "premium")
